feat(Meaning): Split meaning out into languages
Two major reasons for doing this.

1. premature optimisation, now i can support fr, or es meanings.

But more importantly:

2. I'm about to do a bulk download from multiple open dictionary sources. It
occured to me that if I'm doing find or create by (as i wish to) i could end up
with either duplicates or overwriting if i had similar takes on meaning that
are not the same to the character. I'd rather have a list of entries for
meanings and have them be filterable by language.