remove(DictionaryEntry): take pinyin off schema
This was a bad take, it assumes a one-to-one mapping between a character and
it's pronounciation. When I stop to think about it, I know that's not true.
Like the pesky character 张

Also it's a bit limited around other means of prounication, I could wish to add
others in future.

So the current plan is to shift that assumption over to the Meaning, that a
meaning and the pronounciation is intimiately linked, when unique in the
context of a character that likely holds true.

The issue here will be if there is a difference in say Pinyin, but not in
another prounciation system but I'll deal with that if we get to it later